Refactoring: Move import_repository task to galaxy.worker
